add (shell) split cli and the module
